refactor: removed client directory to simplify the project structure
