+++ src/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './client/css/index.css';
-import App from './client/components/App';
-import * as serviceWorker from './client/serviceWorker';
+import './css/index.css';
+import App from './components/App';
+import * as serviceWorker from './serviceWorker';
 import '@statisticsnorway/ssb-component-library/src/main.scss';
 import './i18n';
